Fix: move parallelize package to non-internal file
