+++ pkg/scheduler/framework/plugins/volumerestrictions/volume_restrictions.go
@@ -27,9 +27,9 @@ import (
 	corelisters "k8s.io/client-go/listers/core/v1"
 	v1helper "k8s.io/kubernetes/pkg/apis/core/v1/helper"
 	"k8s.io/kubernetes/pkg/scheduler/framework"
+	"k8s.io/kubernetes/pkg/scheduler/framework/parallelize"
 	"k8s.io/kubernetes/pkg/scheduler/framework/plugins/feature"
 	"k8s.io/kubernetes/pkg/scheduler/framework/plugins/names"
-	"k8s.io/kubernetes/pkg/scheduler/internal/parallelize"
 )
 
 // VolumeRestrictions is a plugin that checks volume restrictions.
